@@ -1,63 +1,69 @@
 {
-	"name": "svelte-template",
-	"version": "0.0.1",
+	"name": "osteopaths-in",
 	"private": true,
+	"version": "0.0.1",
+	"type": "module",
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
+		"prepare": "svelte-kit sync || echo ''",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+		"format": "prettier --write .",
 		"lint": "prettier --check . && eslint .",
-		"format": "prettier --write ."
+		"db:push": "drizzle-kit push",
+		"db:generate": "drizzle-kit generate",
+		"db:migrate": "drizzle-kit migrate",
+		"db:studio": "drizzle-kit studio",
+		"generate-icon": "node scripts/generate-icon/index.mjs"
 	},
 	"devDependencies": {
-		"@epic-web/remember": "^1.0.2",
-		"@lucia-auth/adapter-sqlite": "^3.0.0-beta.13",
-		"@radix-ui/colors": "^3.0.0",
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/adapter-static": "^3.0.1",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@tailwindcss/forms": "^0.5.7",
-		"@tailwindcss/typography": "^0.5.10",
-		"@types/eslint": "8.56.0",
-		"@typescript-eslint/eslint-plugin": "^6.0.0",
-		"@typescript-eslint/parser": "^6.0.0",
-		"arctic": "^1.0.0",
-		"autoprefixer": "^10.4.16",
-		"clsx": "^2.1.0",
-		"drizzle-kit": "^0.20.13",
-		"eslint": "^8.56.0",
-		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
-		"hex-rgb": "^5.0.0",
-		"lucia": "^3.0.0-beta.15",
-		"lucide-svelte": "^0.316.0",
-		"oslo": "^1.0.1",
-		"postcss": "^8.4.32",
-		"postcss-load-config": "^5.0.2",
-		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
-		"prettier-plugin-tailwindcss": "^0.5.9",
-		"radix-icons-svelte": "^1.2.1",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
-		"svelte-markdoc-preprocess": "^2.0.0-rc.4",
-		"tailwind-merge": "^2.2.1",
-		"tailwind-variants": "^0.1.20",
-		"tailwindcss": "^3.3.6",
-		"tslib": "^2.4.1",
+		"@eslint/compat": "^1.2.5",
+		"@eslint/js": "^9.18.0",
+		"@iconify-json/fluent": "^1.2.14",
+		"@iconify-json/logos": "^1.2.4",
+		"@sveltejs/adapter-auto": "^4.0.0",
+		"@sveltejs/kit": "^2.16.0",
+		"@sveltejs/vite-plugin-svelte": "^5.0.0",
+		"@tailwindcss/vite": "^4.0.0",
+		"@types/web-push": "^3.6.4",
+		"drizzle-kit": "^0.30.2",
+		"eslint": "^9.18.0",
+		"eslint-config-prettier": "^10.0.1",
+		"eslint-plugin-svelte": "^2.46.1",
+		"globals": "^15.14.0",
+		"prettier": "^3.4.2",
+		"prettier-plugin-svelte": "^3.3.3",
+		"prettier-plugin-tailwindcss": "^0.6.11",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
+		"svelte-markdoc-preprocess": "^2.1.0",
+		"sveltekit-superforms": "^2.23.1",
+		"tailwindcss": "^4.0.0",
 		"typescript": "^5.0.0",
-		"vite": "^5.0.3"
+		"typescript-eslint": "^8.20.0",
+		"unplugin-icons": "^22.1.0",
+		"vite": "^6.0.0",
+		"zod": "^3.24.2"
 	},
-	"type": "module",
 	"dependencies": {
-		"@libsql/client": "^0.3.6",
-		"@upstash/redis": "^1.28.2",
-		"bits-ui": "^0.15.1",
-		"dotenv": "^16.4.1",
-		"drizzle-orm": "^0.29.3",
-		"svelte-sonner": "^0.3.21"
+		"@fontsource-variable/manrope": "^5.1.1",
+		"@libsql/client": "^0.14.0",
+		"@oslojs/crypto": "^1.0.1",
+		"@oslojs/encoding": "^1.1.0",
+		"@sveltejs/adapter-cloudflare": "^5.0.3",
+		"@vite-pwa/assets-generator": "^0.2.6",
+		"@vite-pwa/sveltekit": "^0.6.7",
+		"arctic": "^3.4.0",
+		"bits-ui": "^1.3.4",
+		"cloudinary": "^2.5.1",
+		"drizzle-orm": "^0.38.4",
+		"mode-watcher": "^0.5.1",
+		"svelte-easy-crop": "^4.0.0",
+		"tailwind-merge": "^3.0.2",
+		"tailwind-variants": "^0.3.1",
+		"vaul-svelte": "^1.0.0-next.6",
+		"web-push": "^3.6.7"
 	}
 }
