@@ -1,5 +1,9 @@
 <script lang="ts">
-	export let size: number | string = 24;
+	type Props = {
+		class?: string;
+		size?: number | string;
+	};
+	let { size, class: className }: Props = $props();
 </script>
 
 <svg
@@ -9,33 +13,33 @@
 	viewBox="0 0 125 125"
 	fill="none"
 	xmlns="http://www.w3.org/2000/svg"
-	class={`lucide-logo animate-spin [animation-duration:60s] ${$$props.class ?? ''}`}
+	class="lucide-logo animate-[spin_60s_linear_infinite;] {className}"
 >
 	<path
 		d="M63.1067 62.0837L10.8749 62.7007"
 		stroke="#22D3EE"
-		style="stroke:#22D3EE;stroke:color(display-p3 0.1333 0.8275 0.9333);stroke-opacity:1;"
+		style="stroke:color(display-p3 0.1333 0.8275 0.9333);stroke-opacity:1;"
 		stroke-width="16"
 		stroke-linecap="round"
 	/>
 	<path
 		d="M63.1066 62.0837L18.9181 33.0657"
 		stroke="#3B82F6"
-		style="stroke:#3B82F6;stroke:color(display-p3 0.2314 0.5098 0.9647);stroke-opacity:1;"
+		style="stroke:color(display-p3 0.2314 0.5098 0.9647);stroke-opacity:1;"
 		stroke-width="16"
 		stroke-linecap="round"
 	/>
 	<path
 		d="M63.1066 62.0837L61.8034 10.0778"
 		stroke="#6366F1"
-		style="stroke:#6366F1;stroke:color(display-p3 0.3882 0.4000 0.9451);stroke-opacity:1;"
+		style="stroke:color(display-p3 0.3882 0.4000 0.9451);stroke-opacity:1;"
 		stroke-width="16"
 		stroke-linecap="round"
 	/>
 	<path
 		d="M63.1064 62.0837L91.5286 19.6398"
 		stroke="#8B5CF6"
-		style="stroke:#8B5CF6;stroke:color(display-p3 0.5451 0.3608 0.9647);stroke-opacity:1;"
+		style="stroke:color(display-p3 0.5451 0.3608 0.9647);stroke-opacity:1;"
 		stroke-width="16"
 		stroke-linecap="round"
 	/>
@@ -43,7 +47,7 @@
 		d="M63.1066 62.0837L33.7972 104.746"
 		stroke="#F97316"
 		stroke-opacity="0.64"
-		style="stroke:#F97316;stroke:color(display-p3 0.9765 0.4510 0.0863);stroke-opacity:0.64;"
+		style="stroke:color(display-p3 0.9765 0.4510 0.0863);stroke-opacity:0.64;"
 		stroke-width="16"
 		stroke-linecap="round"
 	/>
@@ -51,7 +55,7 @@
 		d="M63.1066 62.0837L62.8635 114.65"
 		stroke="#F43F5E"
 		stroke-opacity="0.56"
-		style="stroke:#F43F5E;stroke:color(display-p3 0.9569 0.2471 0.3686);stroke-opacity:0.56;"
+		style="stroke:color(display-p3 0.9569 0.2471 0.3686);stroke-opacity:0.56;"
 		stroke-width="16"
 		stroke-linecap="round"
 	/>
@@ -59,14 +63,14 @@
 		d="M63.1067 62.0837L105.903 91.9512"
 		stroke="#EC4899"
 		stroke-opacity="0.5"
-		style="stroke:#EC4899;stroke:color(display-p3 0.9255 0.2824 0.6000);stroke-opacity:0.5;"
+		style="stroke:color(display-p3 0.9255 0.2824 0.6000);stroke-opacity:0.5;"
 		stroke-width="16"
 		stroke-linecap="round"
 	/>
 	<path
 		d="M63.1066 62.0837L114.597 61.9606"
 		stroke="#A855F7"
-		style="stroke:#A855F7;stroke:color(display-p3 0.6588 0.3333 0.9686);stroke-opacity:1;"
+		style="stroke:color(display-p3 0.6588 0.3333 0.9686);stroke-opacity:1;"
 		stroke-width="16"
 		stroke-linecap="round"
 	/>
@@ -78,6 +82,6 @@
 		rx="7"
 		transform="rotate(-28.2161 53.6289 59.225)"
 		fill="white"
-		style="fill:white;fill:white;fill-opacity:1;"
+		style="fill:white;fill-opacity:1;"
 	/>
 </svg>
