@@ -1,15 +1,17 @@
-// See https://kit.svelte.dev/docs/types#app
+import "vite-plugin-pwa/svelte";
+import "vite-plugin-pwa/info";
+import "vite-plugin-pwa/pwa-assets";
+
+import "unplugin-icons/types/svelte";
+
+// See https://svelte.dev/docs/kit/types#app.d.ts
 // for information about these interfaces
 declare global {
 	namespace App {
-		// interface Error {}
 		interface Locals {
-			user: import('lucia').User | null;
-			session: import('lucia').Session | null;
+			user: import("$lib/server/auth/session").SessionValidationResult["user"];
+			session: import("$lib/server/auth/session").SessionValidationResult["session"];
 		}
-		// interface PageData {}
-		// interface PageState {}
-		// interface Platform {}
 	}
 }
 
