@@ -1,30 +1,18 @@
-import adapter from '@sveltejs/adapter-auto';
-import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
-import { markdoc } from 'svelte-markdoc-preprocess';
-import { dirname, join } from 'path';
-import { fileURLToPath } from 'url';
-
-function absoulute(file) {
-	return join(dirname(fileURLToPath(import.meta.url)), file);
-}
+import adapter from "@sveltejs/adapter-cloudflare";
+import { vitePreprocess } from "@sveltejs/vite-plugin-svelte";
+import { markdoc } from "svelte-markdoc-preprocess";
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-	// Consult https://kit.svelte.dev/docs/integrations#preprocessors
+	// Consult https://svelte.dev/docs/kit/integrations
 	// for more information about preprocessors
-	preprocess: [
-		vitePreprocess({}),
-		markdoc({
-			layouts: {
-				default: absoulute('./src/lib/layouts/Default.svelte')
-			}
-		})
-	],
-	extensions: ['.markdoc', '.svelte'],
+	preprocess: [vitePreprocess(), markdoc()],
+	extensions: [".markdoc", ".svelte"],
+
 	kit: {
-		// adapter-auto only supports some environments, see https://kit.svelte.dev/docs/adapter-auto for a list.
-		// If your environment is not supported or you settled on a specific environment, switch out the adapter.
-		// See https://kit.svelte.dev/docs/adapters for more information about adapters.
+		// adapter-auto only supports some environments, see https://svelte.dev/docs/kit/adapter-auto for a list.
+		// If your environment is not supported, or you settled on a specific environment, switch out the adapter.
+		// See https://svelte.dev/docs/kit/adapters for more information about adapters.
 		adapter: adapter()
 	}
 };
